+++ MergeSort Project 2/generate_random.hpp
@@ -7,14 +7,14 @@
 #include <set>
 using namespace std;
 
-void print_vector(int size, const int* a);
+void print_vector(long size, const long* a);
 
-void write_vector(int size, const int* a);
+void write_vector(long size, const long* a);
 
-void read_vector(int size, int* a);
+void read_vector(long size, long* a);
 
-void generate_vector(int size, int* a, unsigned seed);
+void generate_vector(long size, long* a, unsigned seed);
 
-void generate_norepeat_vector(int size, int* a, unsigned seed);
+void generate_norepeat_vector(long size, long* a, unsigned seed);
 
 #endif
